+++ CHANGELOG.md
@@ -2,4 +2,16 @@
 
 History of **rest2go** library with changes description.
 
-## Next release
+## 1.0.0 (2025-11-03)
+
+- `FEATURE` Health check
+- `FEATURE` Filtering
+- `FEATURE` Pagination
+- `FEATURE` Database migrations by Goose
+- `FEATURE` SQLite and Postgres database connection
+- `FEATURE` Preconfigured HTTP server
+- `FEATURE` Api-Key header authorization middleware
+- `FEATURE` Log HTTP request & response middleware
+- `FEATURE` REST API errors handling
+- `FEATURE` Loaded once settings as YAML file
+- `FEATURE` GoLang library project initialization